Move dependency handling to common
